Renamed util and removed tests of old old util
