+++ src/client/Utils/parseStringForVideoUrls.js
@@ -1,28 +1,9 @@
 /**
- * @file Extracts the youtube video ID from a given string
+ * @file Extracts the video URL from a given string
  */
 
 import React from 'react';
-import {isEmpty} from 'lodash';
-import youtubeIdGetter from 'youtube-link-to-id';
 import ReactPlayer from 'react-player';
-
-/**
- * Extracts the youtube video ID from the given string.
- * If an empty string is given or no IDs are found null is returned. Otherwise
- * an array of ids is returned.
- *
- * @param {String} str
- * @return {Array|null}
- */
-export default function extractYoutubeID(str) {
-  if (isEmpty(str)) {
-    return null;
-  }
-
-  const ids = youtubeIdGetter.linkStringToIds(str);
-  return !isEmpty(ids) ? ids : null;
-}
 
 /**
  * Extracts youtube and vimeo URLs from the given string and return the matches. If returnPlayer is true a player based
