+++ src/client/components/Review/Review.component.js
@@ -5,7 +5,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import TimeToString from '../../Utils/timeToString.js';
-import {parseStringForVideoUrls} from '../../Utils/extractYoutubeID';
+import {parseStringForVideoUrls} from '../../Utils/parseStringForVideoUrls';
 import isSiteOpen from '../../Utils/openingHours';
 
 import Message from '../General/Message/Message.component.js';
